+++ VIStk/Widgets/_MenuItem.py
@@ -3,7 +3,7 @@ from tkinter import ttk
 import os
 import sys
 
-class MenuItem():
+class MenuItem(Button):
     """Each item in the menu is created from the corresponding .json file. Each path should be given relative to xyz/WOM/
     """
     def __init__(self,parent:Frame|Toplevel|LabelFrame|Tk,path,nav,*args,**kwargs):
@@ -14,16 +14,15 @@ class MenuItem():
             path (str): Name of .exe or absolute path to python script
             nav (str): Navigation character to click button
         """
-        self.button = Button(master=parent, *args, **kwargs)
-        self.parent = parent
+        super().__init__(master=parent, *args, **kwargs)
         self.root = parent.winfo_toplevel()
         self.path = path
         self.nav = nav
-        self.button.config(command = self.itemPath)
+        self.config(command = self.itemPath)
         enter = lambda event: event.widget.configure(background="dodger blue")
         leave = lambda event: event.widget.configure(background="gray94")
-        self.button.bind("<Enter>", enter)
-        self.button.bind("<Leave>", leave)
+        self.bind("<Enter>", enter)
+        self.bind("<Leave>", leave)
         #self.button.pack()
 
     def itemPath(self):